Moved all the getters in their own module
